fix: Add test dependencies and fix schema/test issues for #29
- Add shoulda-matchers gem for model validation testing
- Configure factory_bot and shoulda-matchers in rails_helper
- Fix migration spec type expectation (bigint reported as integer by Rails 8.x)
- Fix Rubocop style violations in migration and spec files
- Regenerate schema.rb after migration execution

+++ spec/migrations/20260827175041_create_events_spec.rb
@@ -15,7 +15,7 @@ RSpec.describe 'CreateEvents' do
       expect(columns_hash['title'].type).to eq :string
       expect(columns_hash['description'].type).to eq :text
       expect(columns_hash['event_date'].type).to eq :date
-      expect(columns_hash['created_by_id'].type).to eq :bigint
+      expect(columns_hash['created_by_id'].type).to eq :integer
       expect(columns_hash['spreadsheet_id'].type).to eq :string
       expect(columns_hash['discarded_at'].type).to eq :datetime
     end
@@ -43,7 +43,7 @@ RSpec.describe 'CreateEvents' do
       composite_index = indexes.find { |idx| idx.name == 'index_events_on_event_date_and_discarded_at' }
 
       expect(composite_index).not_to be_nil
-      expect(composite_index.columns).to eq(['event_date', 'discarded_at'])
+      expect(composite_index.columns).to eq([ 'event_date', 'discarded_at' ])
     end
   end
 end
